src/app/modules/public/views/login/login.component.ts
import {Component} from '@angular/core';
import {FormBuilder, Validators} from '@angular/forms';

@Component({
  selector: 'app-login',
  templateUrl: './login.component.html',
  styleUrls: ['./login.component.scss']
})
export class LoginComponent {
  isSubmitted: boolean = false;
  loginForm = this._formBuilder.group({
    email: ['', [Validators.email, Validators.required]],
    password: ['', [Validators.required]]
  });

  constructor(private _formBuilder: FormBuilder) {
  }

  public onLogin(): void {
    this.isSubmitted = true;
  }
}
